UDP: Don't keep retrying send if we get EINVAL
UDP can return EINVAL if the route is down or not configured, this can
cause knet to 'spin' (well, ish, there is a usleep in there).

With this patch in place knet doesn't hog the CPU anywhere near so much
(and it does recover if the link returns).

TBH I'm in 2 minds about this. When I first spotted corosync taking up
100% CPU because I had forgotten to ifup the link I screamed 'bug', but
I'm not totally sure. It might need more sensitive handling. hence I'm posting it here as much for discussion as anything else

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>