JUST A TEST Add dynamic headers to the datafd comms with the application.
Used to pass back IPaddres to corosync and also for corosync
to pass the target NODEID in the packet.

This was all pointless becase
a) The IP address is meaningless if a packet has been fragmented
b) The knet filter works fine, and is needed anyway for RX packets

There might be other useful bits of code in here though so I don't
want to just throw it away!

CC 8-Nov-2024