Add optional header to messages sent back to the application
If enabled in (the new API call) knet_handle_add_datafd_new(),
knet will add a header to the message with extra details. Currently
this is just the nodeid, but other things could, concievably, be
added as the structure is extensible.

This allows the receiving application to double-check that the message
has come from the right node - and also means that it doesn't need to
be added to the application's message header.